add channel/bitset package
