+++ shardingsphere-scaling/shardingsphere-scaling-core/src/main/java/org/apache/shardingsphere/scaling/core/execute/executor/channel/bitset/BitSetChannel.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.scaling.core.execute.executor.channel;
+package org.apache.shardingsphere.scaling.core.execute.executor.channel.bitset;
 
 import org.apache.shardingsphere.scaling.core.execute.executor.record.Record;
